Remove country input wrapper
