+++ vgscheckout/src/androidTest/java/com/verygoodsecurity/vgscheckout/custom/ui/AddressDialogTest.kt
@@ -14,7 +14,7 @@ import androidx.test.espresso.matcher.ViewMatchers.*
 import androidx.test.ext.junit.runners.AndroidJUnit4
 import com.verygoodsecurity.vgscheckout.BuildConfig
 import com.verygoodsecurity.vgscheckout.R
-import com.verygoodsecurity.vgscheckout.collect.widget.VGSCountryEditText
+import com.verygoodsecurity.vgscheckout.collect.view.internal.CountryInputField
 import com.verygoodsecurity.vgscheckout.config.VGSCheckoutCustomConfig
 import com.verygoodsecurity.vgscheckout.config.ui.VGSCheckoutCustomFormConfig
 import com.verygoodsecurity.vgscheckout.config.ui.view.address.VGSCheckoutBillingAddressVisibility
@@ -112,7 +112,7 @@ class AddressDialogTest {
         launch<CustomSaveCardActivity>(intent).use {
             // Act
             var countries: List<Country>? = null
-            onView(withId(R.id.vgsEtCountry)).perform(ActionHelper.doAction<VGSCountryEditText> {
+            onView(withId(R.id.vgsEtCountry)).perform(ActionHelper.doAction<CountryInputField> {
                 countries = it.getCountries()
             })
             //Assert
@@ -129,7 +129,7 @@ class AddressDialogTest {
         launch<CustomSaveCardActivity>(intent).use {
             // Act
             var countries: List<Country>? = null
-            onView(withId(R.id.vgsEtCountry)).perform(ActionHelper.doAction<VGSCountryEditText> {
+            onView(withId(R.id.vgsEtCountry)).perform(ActionHelper.doAction<CountryInputField> {
                 countries = it.getCountries()
             })
             //Assert
@@ -146,7 +146,7 @@ class AddressDialogTest {
         launch<CustomSaveCardActivity>(intent).use {
             // Act
             var countries: List<Country>? = null
-            onView(withId(R.id.vgsEtCountry)).perform(ActionHelper.doAction<VGSCountryEditText> {
+            onView(withId(R.id.vgsEtCountry)).perform(ActionHelper.doAction<CountryInputField> {
                 countries = it.getCountries()
             })
             //Assert
@@ -163,7 +163,7 @@ class AddressDialogTest {
         launch<CustomSaveCardActivity>(intent).use {
             // Act
             var countries: List<Country>? = null
-            onView(withId(R.id.vgsEtCountry)).perform(ActionHelper.doAction<VGSCountryEditText> {
+            onView(withId(R.id.vgsEtCountry)).perform(ActionHelper.doAction<CountryInputField> {
                 countries = it.getCountries()
             })
             //Assert
